+++ src/CHANGELOG.md
@@ -3,6 +3,9 @@
 - A `CHANGELOG.md`. Yay! (This is technically release notes but oh well).
 - This mod's seeds and gourd bottles are now added into villager house loot tables
 - Bottle gourd patches now spawn as a vegetation feature in vanilla savanna biomes, in a similar manner to pumpkins.
+- New structures are added to villages
+  - A small pumpkin garden can now be found in taiga villages
+  - A drinking well can now be found in desert villages
 
 ## Updates
 - Gourd bottles now can be emptied in the crafting table.